requirements.txt
kivy
https://github.com/kivymd/KivyMD/archive/master.zip
matplotlib
https://github.com/kivy-garden/graph/archive/master.zip
kivy_garden.matplotlib
setuptools
plyer
re
